first conv selection, update warnings
