+++ nncf/common/pruning/utils.py
@@ -136,29 +136,6 @@ def traverse_function(node: NNCFNode, output: List[NNCFNode], type_check_fn, vis
     return True, output
 
 
-def get_first_nodes_of_type(graph: NNCFGraph, op_types: List[str]) -> List[NNCFNode]:
-    """
-    Looking for first node in graph with type in `op_types`.
-    First == layer with type in `op_types`, that there is a path from the input such that there are no other
-    operations with type in `op_types` on it.
-
-    :param op_types: Types of modules to track.
-    :param graph: Graph to work with.
-    :return: List of all first nodes with type in `op_types`.
-    """
-    graph_roots = graph.get_input_nodes()  # NNCFNodes here
-
-    visited = {node_id: False for node_id in graph.get_all_node_ids()}
-    partial_traverse_function = partial(traverse_function,
-                                        type_check_fn=lambda x: x in op_types,
-                                        visited=visited)
-
-    first_nodes_of_type = []
-    for root in graph_roots:
-        first_nodes_of_type.extend(graph.traverse_graph(root, partial_traverse_function))
-    return first_nodes_of_type
-
-
 def get_last_nodes_of_type(graph: NNCFGraph, op_types: List[str]) -> List[NNCFNode]:
     """
     Looking for last node in graph with type in `op_types`.
